merge

+++ TMessagesProj/jni/boringssl/include/openssl/cast.h
@@ -82,11 +82,11 @@ OPENSSL_EXPORT void CAST_ecb_encrypt(const uint8_t *in, uint8_t *out,
 OPENSSL_EXPORT void CAST_encrypt(uint32_t *data, const CAST_KEY *key);
 OPENSSL_EXPORT void CAST_decrypt(uint32_t *data, const CAST_KEY *key);
 OPENSSL_EXPORT void CAST_cbc_encrypt(const uint8_t *in, uint8_t *out,
-                                     long length, const CAST_KEY *ks,
+                                     size_t length, const CAST_KEY *ks,
                                      uint8_t *iv, int enc);
 
 OPENSSL_EXPORT void CAST_cfb64_encrypt(const uint8_t *in, uint8_t *out,
-                                       long length, const CAST_KEY *schedule,
+                                       size_t length, const CAST_KEY *schedule,
                                        uint8_t *ivec, int *num, int enc);
 
 #ifdef  __cplusplus
